refactor: streamline GitCue extension by consolidating services and enhancing dashboard interactions
- Removed unused interfaces and consolidated functionality into dedicated services for better maintainability.
- Updated status bar and dashboard to reflect real-time changes using the new service architecture.
- Improved command handling for toggling watch status and committing with preview.
- Enhanced error handling and logging for better debugging and user experience.
