+++ gitcue/src/services/index.ts
@@ -1,0 +1,5 @@
+export { GitCueStatusProvider } from './statusProvider';
+export { DashboardService } from './dashboardService';
+export { ActivityLogger } from './activityLogger';
+export { CommitService } from './commitService';
+export { FileWatcherService } from './fileWatcherService';
